DeviceAdapter compiler warning removal
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@17201 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/PrecisExcite/PrecisExcite.cpp
@@ -470,9 +470,9 @@ void Controller::Illuminate()
    {
       if (triggerMode_ == OFF) {
          msg << "SQZ" << carriage_return;
-         for (int i=0; i<channelLetters_.size(); i++) {
+         for (unsigned int i=0; i<channelLetters_.size(); i++) {
             msg << "C" << channelLetters_[i];
-            if (i == currentChannel_)
+            if (i == (unsigned int)currentChannel_)
                msg << "N";
             else
                msg << "F";
